update: checks if member has a pending book to return before being deleted

+++ app/members/views.py
@@ -2,17 +2,16 @@ from app import db
 from app.members import members_bp
 from flask import request, jsonify
 from ..schema import MemberSchema
-from ..models import Member
+from ..models import Member, Transaction
+from ..utils import TransactionType
 
 from math import ceil
 from pydantic import ValidationError
 
-members_error_dict = {
-    'Error': 'Could not find member!'
-}
+members_error_dict = {"Error": "Could not find member!"}
 
 
-@members_bp.route('/hello')
+@members_bp.route("/hello")
 def hello():
     """Returns a string.
 
@@ -22,7 +21,7 @@ def hello():
     return "Hello from the members side"
 
 
-@members_bp.route('/create', methods=['POST'])
+@members_bp.route("/create", methods=["POST"])
 def create_member():
     """Creates a member object and saves to database.
 
@@ -35,27 +34,29 @@ def create_member():
         member_schema = MemberSchema(**member)
 
         member = Member(
-            name = member_schema.name,
-            debt = member_schema.debt,
-            books_borrowed = member_schema.books_borrowed
+            name=member_schema.name,
+            debt=member_schema.debt,
+            books_borrowed=member_schema.books_borrowed,
         )
 
         db.session.add(member)
         db.session.commit()
 
-        return jsonify({
-            'Message': 'Member created succesfully',
-            'Member': member_schema.model_dump()
-        }), 201
+        return (
+            jsonify(
+                {
+                    "Message": "Member created succesfully",
+                    "Member": member_schema.model_dump(),
+                }
+            ),
+            201,
+        )
 
     except ValidationError as e:
-        return jsonify({
-            'Error': 'Validation failed',
-            'Details': e.errors()
-        }), 400
+        return jsonify({"Error": "Validation failed", "Details": e.errors()}), 400
 
 
-@members_bp.route('/update/<int:member_id>', methods=['PUT'])
+@members_bp.route("/update/<int:member_id>", methods=["PUT"])
 def update_member(member_id):
     """Updates a member object.
 
@@ -78,24 +79,22 @@ def update_member(member_id):
 
         member.name = member_schema.name
         member.debt = member_schema.debt
-        member.books_borrowed =  member_schema.books_borrowed
+        member.books_borrowed = member_schema.books_borrowed
 
         db.session.commit()
 
-        return jsonify({
-            'Message': "Member updated successfully!",
-            'Member': member_schema.model_dump()
-        })
-
+        return jsonify(
+            {
+                "Message": "Member updated successfully!",
+                "Member": member_schema.model_dump(),
+            }
+        )
 
     except ValidationError as e:
-        return jsonify({
-            'Error': 'Validation failed',
-            'Details': e.errors()
-        }), 400
+        return jsonify({"Error": "Validation failed", "Details": e.errors()}), 400
 
 
-@members_bp.route('/get_by_id/<int:member_id>')
+@members_bp.route("/get_by_id/<int:member_id>")
 def get_member_by_id(member_id):
     """Get member object using id.
 
@@ -111,49 +110,58 @@ def get_member_by_id(member_id):
     if member is None:
         return members_error_dict, 400
 
-    return jsonify({
-        'id': member.id,
-        'name': member.name,
-        'debt': member.debt,
-        'books_borrowed':member.books_borrowed
-    }), 200
+    return (
+        jsonify(
+            {
+                "id": member.id,
+                "name": member.name,
+                "debt": member.debt,
+                "books_borrowed": member.books_borrowed,
+            }
+        ),
+        200,
+    )
 
 
-@members_bp.route('/get_members')
+@members_bp.route("/get_members")
 def get_members():
     """Gets members in pages.
 
     Returns:
         dict: Pagination object with data as list.
     """
-    page = request.args.get('page', default=1, type=int)
-    per_page = request.args.get('per_page', default=10, type=int)
+    page = request.args.get("page", default=1, type=int)
+    per_page = request.args.get("per_page", default=10, type=int)
 
     members = Member.query.paginate(page=page, per_page=per_page)
 
     total_pages = ceil(members.total / per_page)
 
-    members_list = [{
-        'id': member.id,
-        'name': member.name,
-        'debt': member.debt
-    } for member in members]
+    members_list = [
+        {"id": member.id, "name": member.name, "debt": member.debt}
+        for member in members
+    ]
 
-    return jsonify({
-        'total_members': members.total,
-        'total_pages': total_pages,
-        'pages': members.pages,
-        'current_page': page,
-        'members': members_list,
-        'per_page': per_page,
-        'has_next': members.has_next,
-        'has_prev': members.has_prev,
-        'next_page': members.next_num if members.has_next else None,
-        'prev_page': members.prev_num if members.has_prev else None
-    }), 200
+    return (
+        jsonify(
+            {
+                "total_members": members.total,
+                "total_pages": total_pages,
+                "pages": members.pages,
+                "current_page": page,
+                "members": members_list,
+                "per_page": per_page,
+                "has_next": members.has_next,
+                "has_prev": members.has_prev,
+                "next_page": members.next_num if members.has_next else None,
+                "prev_page": members.prev_num if members.has_prev else None,
+            }
+        ),
+        200,
+    )
 
 
-@members_bp.route('/delete/<int:member_id>', methods=['DELETE'])
+@members_bp.route("/delete/<int:member_id>", methods=["DELETE"])
 def delete_member(member_id):
     """Deletes a member from the database.
 
@@ -169,9 +177,14 @@ def delete_member(member_id):
     if member is None:
         return members_error_dict, 400
 
+    pending_transactions = Transaction.query.filter_by(
+        member_id=member_id, type=TransactionType.ISSUE
+    ).first()
+
+    if pending_transactions is not None:
+        return jsonify({"Error": "Member has pending transactions!"}), 400
+
     db.session.delete(member)
     db.session.commit()
 
-    return jsonify({
-        'Message': 'Member deleted succesfully!'
-    }), 200
+    return jsonify({"Message": "Member deleted succesfully!"}), 200
